feat!: styles are now fetched locally from the given file by default
Before this change, styles were fetched from the file's library, which first had to be published. Publishing styles to a library is not possible for free Figma accounts. The `stylesFromLibrary` config parameter was added and is `false` by default. Set it to true, to restore the previous behavior and fetch styles from the Library again.